feat: Added logic to check if new toolchain needed
This change introduces a "toolchain_version" file that will be checked
when downloading a toolchain. The child role will be the role that
creates this file. This role will check if that file exists and if it
does, compares the version with the inputted version variable. If the
versions differ, the toolchain will be downloaded.
If the file does not exist, the toolchain will be downloaded.
If the file does exist and the versions match, this role will pass.
Also, this role will now always pass the name of the directory that is
extracted from the tarball.